requirements.txt
anyio==4.9.0
APScheduler==3.10.4
asttokens==3.0.0
catboost==1.2.8
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
faiss-cpu==1.10.0
feedparser==6.0.11
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.8
httpx==0.26.0
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==20.8
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.2
setuptools==78.1.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wcwidth==0.2.13
xgboost==3.0.0
